+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.xx-dev
 Release date: ??.??.????
 
+* Updated styleguide (1.7.0 -> 1.13.0) (HUB-412)
+* Degree programme specific news feed (HUB-378)
 * Added a chat that can be displayed on specific pages (HUB-410)